# feat(CRA-240): Complete FinOps Cost Tracking & Optimization Engine - 100% Test Pass Rate