--- a/services/finops_engine/__init__.py
+++ b/services/finops_engine/__init__.py
@@ -1,0 +1,11 @@
+"""
+FinOps Engine - Real-time Cost Data Collection Engine (CRA-240)
+
+Tracks costs for:
+- OpenAI API calls with token-level granularity
+- Kubernetes resources
+- Vector DB queries
+- Monitoring infrastructure
+
+Target: $0.02 per post cost with sub-second latency storage.
+"""