… .workflows/active_epics.json
… .workflows/epics/epic-240-finops-cost-tracking-optimization.yaml
… .workflows/feature_registry.json
… .workflows/features/feat-240-001-cost-data-collection.yaml
… .workflows/features/feat-240-002-per-post-attribution.yaml
… .workflows/features/feat-240-003-anomaly-detection.yaml
… .workflows/features/feat-240-004-optimization-recommendations.yaml
… .workflows/features/feat-240-005-roi-calculator.yaml
… .workflows/features/feat-240-006-cost-dashboards.yaml
… .workflows/features/feat-240-007-budget-controls.yaml
… .workflows/features/feat-240-008-reporting-api.yaml
… FINOPS_ENGINE_TECHNICAL_DOCUMENTATION.md
… FINOPS_PERFORMANCE_OPTIMIZATION_REPORT.md
… services/common/tests/test_cicd_pipeline_comprehensive_integration.py
… services/common/tests/test_cicd_pipeline_integration.py
… services/common/tests/test_github_workflows.py
… services/common/tests/test_gradual_rollout_manager.py
… services/common/tests/test_gradual_rollout_manager_advanced.py
… services/common/tests/test_gradual_rollout_manager_failure_scenarios.py
… services/common/tests/test_performance_regression_detector.py
… services/common/tests/test_performance_regression_detector_advanced.py
… services/common/tests/test_rollback_controller.py
… services/common/tests/test_rollback_controller_core.py
… services/finops_engine/cost_attribution_api.py
… services/finops_engine/cost_event_storage.py
… services/finops_engine/demo_finops.py
… services/finops_engine/infrastructure_cost_tracker.py
… services/finops_engine/kubernetes_deployment_optimized.yaml
… services/finops_engine/main.py
… services/finops_engine/models.py